generate site index list dynamically
This _navigation.json file is used by the Federalist platform to list
individual files that should be editable in the content editor. Instead
of requiring a content manager to create a file *and* add an entry in a
separate .yml file in the _data directory, we should just pull these
dynamically.

https://federalist-docs.18f.gov/pages/custom-templates/#navigation-menu