build product submission modal
The modal is derived from vote.usa.gov. The structure of the modal is
in place, but the content is pulled from a sample vote.gov selection.
The modal is subscribed to anything with the js-load-submission class.
For now, this is the ‘submit product’ buttons on the header and the
hero on the homepage.