build liquid template to generate JSON of user/category facets
_facets.json will be a JSON blob visible on the site that contains a
dynamically-created list of pages that are mapped to our different
facets. We can use this from within Javascript in order to determine
which pages should be visible when different facets are selected.

The whitespace on the JSON isn’t ideal, but won’t impact the ability to
parse the data. Ideally we would use the {%- tags within Liquid to
remove the whitespace, but that doesn’t seem to be implemented in
Jekyll yet.