dynamically generate faceted search from list of facets
There was very fragile coupling with the User and Category facets. The
checkboxes needed to be setup specifically with the names repeated as
name and ID fields. They needed to be the same as the _facets.json
file. And they needed to be the same on the product pages.

Now, the user and category lists can be defined by the user in
Federalist, and they will be shared in all necessary places. And this
content will still be generated statically by jekyll before serving.